Add Angular 19 support
