--- angular-workspace/ng15/README.md
@@ -1,13 +1,0 @@
-# AngularWorkspace
-
-This project was generated with [Angular CLI](https://github.com/angular/angular-cli) version 15.2.9.
-
-This Angular Project is a utility wrapper to help build the `@trimble-oss/modus-angular-components` library.
-
-All components in this repository are auto generated by the StencilJS library.
-
-To build a local copy of these components you can run:
-
-`npx ng run @trimble-oss/modus-angular-components:build:production`
-
-This can be consumed in a local angular application using npm link
