--- angular-workspace/ng15/package.json
@@ -1,47 +1,0 @@
-{
-  "name": "angular-workspace",
-  "version": "0.0.0-1",
-  "private": true,
-  "scripts": {
-    "ng": "ng",
-    "start:test-harness": "ng run test-harness:serve",
-    "build": "ng run @trimble-oss/modus-angular-components:build:production",
-    "watch": "ng build --watch --configuration development",
-    "test": "ng test"
-  },
-  "dependencies": {
-    "@angular/animations": "^15.2.9",
-    "@angular/common": "^15.2.9",
-    "@angular/compiler": "^15.2.9",
-    "@angular/core": "^15.2.9",
-    "@angular/forms": "^15.2.9",
-    "@angular/platform-browser": "^15.2.9",
-    "@angular/platform-browser-dynamic": "^15.2.9",
-    "@angular/router": "^15.2.9",
-    "@trimble-oss/modus-web-components": "0.41.1",
-    "rxjs": "~7.8.1",
-    "tslib": "^2.5.3",
-    "zone.js": "~0.12.0"
-  },
-  "devDependencies": {
-    "@angular-devkit/build-angular": "^15.2.9",
-    "@angular/cli": "^15.2.9",
-    "@angular/compiler-cli": "^15.2.9",
-    "@types/jasmine": "~3.10.0",
-    "@types/node": "^12.11.1",
-    "jasmine-core": "~4.0.0",
-    "karma": "~6.4.2",
-    "karma-chrome-launcher": "~3.2.0",
-    "karma-coverage": "~2.2.0",
-    "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "~1.7.0",
-    "ng-packagr": "^15.2.2",
-    "typescript": "~4.8.0"
-  },
-  "engines": {
-    "node": ">=14.20.0"
-  },
-  "volta": {
-    "node": "16.20.2"
-  }
-}
