--- angular-workspace/ng15/projects/test-harness/src/app/app.component.html
@@ -1,4 +1,0 @@
-<h1>Modus Angular Components Test Harness</h1>
-<div>
-  <modus-select [label]="'My Modus Select Element'" [options]="[{ key: 'Option 1', value: 'option_1' }]" optionsDisplayProp="key" (valueChange)="onValueChange()"></modus-select>
-</div>
