--- angular-workspace/ng15/projects/test-harness/src/app/app.component.ts
@@ -1,14 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-root',
-  templateUrl: './app.component.html',
-  styleUrls: ['./app.component.scss'],
-})
-export class AppComponent {
-  title = 'test-harness';
-
-  onValueChange() {
-    console.log('Value Changed');
-  }
-}
