--- angular-workspace/ng15/projects/trimble-oss/modus-angular-components/.browserslistrc
@@ -1,19 +1,0 @@
-# This file is used by the build system to adjust CSS and JS output to support the specified browsers below.
-# For additional information regarding the format and rule options, please see:
-# https://github.com/browserslist/browserslist#queries
-
-# For the full list of supported browsers by the Angular framework, please see:
-# https://angular.io/guide/browser-support
-
-# You can see what browsers were selected by your queries by running:
-#   npx browserslist
-
->1%
-last 3 Chrome versions
-last 2 Firefox versions
-last 2 Edge major versions
-last 3 Safari major versions
-last 3 iOS major versions
-last 3 ChromeAndroid versions
-Firefox ESR
-not dead
