--- angular-workspace/ng15/projects/trimble-oss/modus-angular-components/.npmrc
@@ -1,2 +1,0 @@
-lockfile-version = "2"
-registry = https://registry.npmjs.org/
