--- angular-workspace/ng15/projects/trimble-oss/modus-angular-components/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "@trimble-oss/modus-angular-components",
-  "version": "1.7.0-ng15",
-  "license": "MIT",
-  "description": "Trimble Modus Angular Components Library",
-  "homepage": "https://modus-web-components.trimble.com/",
-  "bugs": {
-    "url": "https://github.com/trimble-oss/modus-web-components/issues/"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/trimble-oss/modus-web-components.git"
-  },
-  "peerDependencies": {
-    "@angular/common": "^15.2.9",
-    "@angular/core": "^15.2.9",
-    "@trimble-oss/modus-web-components": "1.7.0"
-  },
-  "dependencies": {
-    "tslib": "^2.5.3"
-  },
-  "engines": {
-    "node": ">=14.20.0"
-  }
-}
