--- angular-workspace/ng15/projects/trimble-oss/modus-angular-components/src/lib/modus-angular-components.module.ts
@@ -1,9 +1,0 @@
-import { NgModule } from '@angular/core';
-import { DIRECTIVES } from './stencil-generated';
-
-@NgModule({
-  declarations: [...DIRECTIVES],
-  imports: [],
-  exports: [...DIRECTIVES],
-})
-export class ModusAngularComponentsModule {}
