--- angular-workspace/ng15/tsconfig.json
@@ -1,40 +1,0 @@
-{
-  "compileOnSave": false,
-  "compilerOptions": {
-    "paths": {
-      "@trimble-oss/modus-angular-components": [
-        "dist/trimble-oss/modus-angular-components/trimble-oss-modus-angular-components",
-        "dist/trimble-oss/modus-angular-components"
-      ]
-    },
-    "baseUrl": "./",
-    "outDir": "./dist/out-tsc",
-    "forceConsistentCasingInFileNames": true,
-    "strict": true,
-    "noImplicitOverride": true,
-    "noPropertyAccessFromIndexSignature": true,
-    "noImplicitReturns": true,
-    "noFallthroughCasesInSwitch": true,
-    "sourceMap": true,
-    "declaration": false,
-    "downlevelIteration": true,
-    "experimentalDecorators": true,
-    "moduleResolution": "node",
-    "importHelpers": true,
-    "preserveSymlinks": true,
-    "skipLibCheck": true,
-    "target": "ES2022",
-    "module": "es2020",
-    "lib": [
-      "es2020",
-      "dom"
-    ],
-    "useDefineForClassFields": false
-  },
-  "angularCompilerOptions": {
-    "enableI18nLegacyMessageIdFormat": false,
-    "strictInjectionParameters": true,
-    "strictInputAccessModifiers": true,
-    "strictTemplates": true
-  }
-}
