+++ angular-workspace/ng19/package.json
@@ -1,0 +1,46 @@
+{
+  "name": "angular-workspace",
+  "version": "0.0.0-1",
+  "private": true,
+  "scripts": {
+    "ng": "ng",
+    "start:test-harness": "ng run test-harness:serve",
+    "build": "ng run @trimble-oss/modus-angular-components:build:production",
+    "watch": "ng build --watch --configuration development",
+    "test": "ng test"
+  },
+  "dependencies": {
+    "@angular/animations": "^19.2.11",
+    "@angular/common": "^19.2.11",
+    "@angular/compiler": "^19.2.11",
+    "@angular/core": "^19.2.11",
+    "@angular/forms": "^19.2.11",
+    "@angular/platform-browser": "^19.2.11",
+    "@angular/platform-browser-dynamic": "^19.2.11",
+    "@angular/router": "^19.2.11",
+    "@trimble-oss/modus-web-components": "1.7.0",
+    "rxjs": "~7.8.0",
+    "tslib": "^2.3.0",
+    "zone.js": "~0.15.0"
+  },
+  "devDependencies": {
+    "@angular-devkit/build-angular": "^19.2.13",
+    "@angular/cli": "^19.2.13",
+    "@angular/compiler-cli": "^19.2.11",
+    "@types/jasmine": "~5.1.0",
+    "jasmine-core": "~5.1.0",
+    "karma": "~6.4.0",
+    "karma-chrome-launcher": "~3.2.0",
+    "karma-coverage": "~2.2.0",
+    "karma-jasmine": "~5.1.0",
+    "karma-jasmine-html-reporter": "~2.1.0",
+    "ng-packagr": "^19.2.2",
+    "typescript": "~5.5.2"
+  },
+  "engines": {
+    "node": ">=18.19.1"
+  },
+  "volta": {
+    "node": "20.19.2"
+  }
+}
