+++ angular-workspace/ng19/projects/test-harness/src/app/app.component.html
@@ -1,0 +1,8 @@
+<h1>Modus Angular Components Test Harness</h1>
+<div>
+  <modus-select
+    [label]="'My Modus Select Element'"
+    [options]="[{ key: 'Option 1', value: 'option_1' }]"
+    optionsDisplayProp="key"
+    (valueChange)="onValueChange()"></modus-select>
+</div>
