+++ angular-workspace/ng19/projects/test-harness/src/app/app.component.ts
@@ -1,0 +1,16 @@
+import { Component } from '@angular/core';
+import { ModusAngularComponentsModule } from '../../../trimble-oss/modus-angular-components/src/lib/modus-angular-components.module';
+
+@Component({
+    selector: 'app-root',
+    imports: [ModusAngularComponentsModule],
+    templateUrl: './app.component.html',
+    styleUrl: './app.component.scss'
+})
+export class AppComponent {
+  title = 'test-harness';
+
+  onValueChange() {
+    console.log('Value Changed');
+  }
+}
