+++ angular-workspace/ng19/projects/trimble-oss/modus-angular-components/.npmrc
@@ -1,0 +1,2 @@
+lockfile-version = "3"
+registry = https://registry.npmjs.org/
