+++ angular-workspace/ng19/projects/trimble-oss/modus-angular-components/ng-package.json
