+++ angular-workspace/ng19/projects/trimble-oss/modus-angular-components/package.json
@@ -1,0 +1,29 @@
+{
+  "name": "@trimble-oss/modus-angular-components",
+  "version": "1.7.0-ng19",
+  "license": "MIT",
+  "description": "Trimble Modus Angular Components Library",
+  "homepage": "https://modus-web-components.trimble.com/",
+  "bugs": {
+    "url": "https://github.com/trimble-oss/modus-web-components/issues/"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/trimble-oss/modus-web-components.git"
+  },
+  "peerDependencies": {
+    "@angular/common": "^19.2.11",
+    "@angular/core": "^19.2.11",
+    "@trimble-oss/modus-web-components": "1.7.0"
+  },
+  "dependencies": {
+    "tslib": "^2.3.0"
+  },
+  "engines": {
+    "node": ">=18.19.1"
+  },
+  "volta": {
+    "node": "20.19.2"
+  },
+  "sideEffects": false
+}
