+++ angular-workspace/ng19/projects/trimble-oss/modus-angular-components/src/lib/modus-angular-components.module.ts
@@ -1,0 +1,9 @@
+import { NgModule } from '@angular/core';
+import { DIRECTIVES } from './stencil-generated';
+
+@NgModule({
+  declarations: [...DIRECTIVES],
+  imports: [],
+  exports: [...DIRECTIVES],
+})
+export class ModusAngularComponentsModule { }
