+++ package.json
@@ -17,13 +17,13 @@
     "lint-links": "npx linkinator https://modus-web-components.trimble.com/ --recurse",
     "prettier": "npx prettier --write \"**/*.yml\"",
     "spellcheck": "npx cspell \"**/*.{html,json,md,mdx,toml,ts,yml}\" --no-progress",
-    "build-angular": "cd angular-workspace/ng15 && npm i && npm run build && cd ../ng16 && npm i && npm run build && cd ../ng17 && npm i && npm run build && cd ../ng18 && npm i && npm run build",
-    "update-mwc-deps": "cd angular-workspace/ng15/projects/trimble-oss/modus-angular-components && npx npm-check-updates -u --dep peer @trimble-oss* && npm i && cd ../../../../ng16/projects/trimble-oss/modus-angular-components && npx npm-check-updates -u --dep peer @trimble-oss* && npm i && cd ../../../../ng17/projects/trimble-oss/modus-angular-components && npx npm-check-updates -u --dep peer @trimble-oss* && npm i && cd ../../../../ng18/projects/trimble-oss/modus-angular-components && npx npm-check-updates -u --dep peer @trimble-oss* && npm i && cd ../../../../../react-workspace/react-17 && npx npm-check-updates -u @trimble-oss* && npm i && cd ../react-18 && npx npm-check-updates -u @trimble-oss* && npm i && cd ../react-19 && npx npm-check-updates -u @trimble-oss* && npm i"
+    "build-angular": "cd angular-workspace/ng16 && npm i && npm run build && cd ../ng17 && npm i && npm run build && cd ../ng18 && npm i && npm run build && cd ../ng19 && npm i && npm run build",
+    "update-mwc-deps": "cd angular-workspace/ng16/projects/trimble-oss/modus-angular-components && npx npm-check-updates -u --dep peer @trimble-oss* && npm i && cd ../../../../ng17/projects/trimble-oss/modus-angular-components && npx npm-check-updates -u --dep peer @trimble-oss* && npm i && cd ../../../../ng18/projects/trimble-oss/modus-angular-components && npx npm-check-updates -u --dep peer @trimble-oss* && npm i && cd ../../../../ng19/projects/trimble-oss/modus-angular-components && npx npm-check-updates -u --dep peer @trimble-oss* && npm i && cd ../../../../../react-workspace/react-17 && npx npm-check-updates -u @trimble-oss* && npm i && cd ../react-18 && npx npm-check-updates -u @trimble-oss* && npm i && cd ../react-19 && npx npm-check-updates -u @trimble-oss* && npm i"
   },
   "engines": {
     "node": ">=18"
   },
   "volta": {
-    "node": "18.20.8"
+    "node": "20.19.2"
   }
 }
